Python code for stand alone map-reduce solution testing
MapReduce.py does the job of a Name-Node in a Hadoop cluster. This code
can be used to test you map-reduce solution on small datasets for
validation. Some of the sample programs provided can be used to
understand the working. When executing the sample code make sure the
MapReduce.py and the example code are in the same directory.